reconfigure configuration plugin
